Remove the canvas save/restore calls in Paragraph::Paint (#4822)
This is a workaround for https://github.com/flutter/flutter/issues/15702

Skia was not drawing the paragraph's text blobs in some cases when a
save/restore sequence is done around the canvas->translate call.  The
problem does not occur if an offsetting call to translate is made instead
of the save/restore.  Will need to work with Skia to find the root cause.