Revert spelling correction in licence matcher (#7242)
In ce1e18057357c2d167ed7b64a0144220502182e5 (#7179), the word
'publically' was altered to 'publicly', causing the licence aggregator
to fail to match the licence text included in many BoringSSL source
files.

The original licence includes several misspellings (e.g. 'rouines'
rather than 'routines') which we need to faithfully reproduce in our
match regex.

The failure this triggered went unnoticed during pre-/post-submit
due to the fact that the contents of the third_party directory didn't
change, so hash of the list of files used as a signature matched the
signature in the golden file, causing the run to be skipped.